added validation for all endpoints, postman collection

+++ src/main/java/com/apress/dayle/rest/UserRegistrationRestController.java
@@ -1,6 +1,7 @@
 package com.apress.dayle.rest;
 
 import com.apress.dayle.dto.UsersDTO;
+import com.apress.dayle.exception.CustomErrorType;
 import com.apress.dayle.repository.UserJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -30,11 +31,21 @@ public class UserRegistrationRestController {
     @GetMapping("/")
     public ResponseEntity<List<UsersDTO>> listAllUsers() {
         List<UsersDTO> users = userJpaRepository.findAll();
+        if (users.isEmpty()) {
+            return new ResponseEntity<List<UsersDTO>>(HttpStatus.NO_CONTENT);
+        }
         return new ResponseEntity<List<UsersDTO>>(users, HttpStatus.OK);
     }
 
     @PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UsersDTO> createUser(@RequestBody final UsersDTO user) {
+        if (userJpaRepository.findByName(user.getName()) != null) {
+            return new ResponseEntity<UsersDTO>(
+                    new CustomErrorType(
+                            "Unable to create new user. A User with name " + user.getName() + " already exist."),
+                    HttpStatus.CONFLICT);
+        }
+
         userJpaRepository.save(user);
         return new ResponseEntity<UsersDTO>(user, HttpStatus.CREATED);
     }
@@ -42,6 +53,11 @@ public class UserRegistrationRestController {
     @GetMapping("/{id}")
     public ResponseEntity<UsersDTO> getUserById(@PathVariable("id") final Long id) {
         UsersDTO user = userJpaRepository.findById(id);
+        if(user == null) {
+            return new ResponseEntity<UsersDTO>(new CustomErrorType(
+                    "User with id "
+                     + id + " not found."), HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<UsersDTO>(user, HttpStatus.OK);
     }
 
@@ -50,6 +66,11 @@ public class UserRegistrationRestController {
             @PathVariable("id") final Long id, @RequestBody UsersDTO user) {
         // fetch user based on id and set it to currentUser object of type UserDTO
         UsersDTO currentUser = userJpaRepository.findById(id);
+        if (currentUser == null) {
+            return new ResponseEntity<UsersDTO>(
+                    new CustomErrorType("Unable to upate. User with id "
+                            + id + " not found."), HttpStatus.NOT_FOUND);
+        }
         // update currentUser object data with user object data
         currentUser.setName(user.getName());
         currentUser.setAddress(user.getAddress());
@@ -62,7 +83,13 @@ public class UserRegistrationRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<UsersDTO> deleteUser(@PathVariable("id") final Long id) {
+        UsersDTO user = userJpaRepository.findById(id);
+        if (user == null) {
+            return new ResponseEntity<UsersDTO>(
+                    new CustomErrorType("Unable to delete. User with id " + id + " not found."), HttpStatus.NOT_FOUND);
+        }
         userJpaRepository.delete(id);
-        return new ResponseEntity<UsersDTO>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<UsersDTO>(new CustomErrorType("Deleted User with id " + id + "."),
+                HttpStatus.NO_CONTENT);
     }
 }
